// source/web/src/api/generated/model/reportChatResponse.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * AdventureWorksAIWorkspaceAPI.Api
 * OpenAPI spec version: 1.0
 */
import type { ReportDetailsDto } from './reportDetailsDto'
import type { ReportMessageDto } from './reportMessageDto'
import type { ReportChatResponseSqlQuery } from './reportChatResponseSqlQuery'
import type { ReportChatResponseOutcome } from './reportChatResponseOutcome'
import type { ReportChatResponseResult } from './reportChatResponseResult'
import type { ChartSpec } from './chartSpec'

export interface ReportChatResponse {
  report: ReportDetailsDto
  userMessage: ReportMessageDto
  assistantMessage: ReportMessageDto
  /** @nullable */
  sqlQuery?: ReportChatResponseSqlQuery
  outcome: ReportChatResponseOutcome
  /** @nullable */
  message?: string | null
  /** @nullable */
  result?: ReportChatResponseResult
  charts: ChartSpec[]
}