// source/web/src/api/generated/model/reportMessageRole.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * AdventureWorksAIWorkspaceAPI.Api
 * OpenAPI spec version: 1.0
 */

export type ReportMessageRole =
  (typeof ReportMessageRole)[keyof typeof ReportMessageRole]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ReportMessageRole = {
  User: 'User',
  Assistant: 'Assistant',
  System: 'System',
} as const